styling
- Add icons on all menu items, footer menu items also (menu icons guide
attached).
- Add a rollover state on the top menu, nested items in the orange bar.
- change the body font in the carousel from Arial to Open Sans, also
make it a big bigger.  Title font and size are fine.
- The image in the banner justifies left on a wider 1920x1200 monitor,
see screen shot below.
- In the footer menu 'Getting Started' needs to move under Help,
'Forums' moved under Community.
- use different font awesome arrow on the get Started Button.  We can
use that same arrow as the icon for the menu item (see menu icons guide)
- make the arrows of the carousel bigger and about the same grey colour
as the actual carousel and apply shadows so they look more subtle.
- First paragraph of text on the BridgeIt page is a different size and
colour that the remaining body font on that page.
- Similarly, the Get Started body font size and colour is different on
the Help page.