android install #19 don't redirect
redirecting from /android/install works ok if the user opens the app
after install, but not if hitting the back button from the play store
or opening Chrome up manually after that. The page will no longer
redirect, but will still show the user the buttons to install or go
back to their app.